Mooved reviewModel to utils
